\name{experimental}
\alias{as_intuBag}
\alias{intuBag}
\alias{is_intuBag}
\alias{intuEnv}
\alias{set_intuEnv}
\title{Experimental features.}
\description{
  Undocumented experimental features. Functional, but under development, and can
  change until stabilization. Not recommended for production code.}
\usage{
as_intuBag(object)
intuBag(...)
is_intuBag(object)
intuEnv()
set_intuEnv(envir)
}
\arguments{
  \item{object}{Object.}
  \item{envir}{Environment.}
  \item{...}{Other arguments.}
}
\value{
  Return value of functions.
}
\author{
Roberto Bertolusso
}
\examples{
library(intubate)

## None yet
}
\keyword{intubate}
\keyword{magrittr}
