\name{intuBag}
\alias{intuBag}
\alias{is_intuBag}
\title{Experimental features.}
\description{
  Undocumented experimental features.}
\usage{
intuBag(...)
is_intuBag(object)
}
\arguments{
  \item{object}{Object.}
  \item{...}{Other arguments.}
}
\value{
  Return value of functions.
}
\author{
Roberto Bertolusso
}
\examples{
library(intubate)

## None yet
}
\keyword{intubate}
\keyword{magrittr}
