\name{experimental}
\alias{as_intuBag}
\alias{clear_intuEnv}
\alias{intuBag}
\alias{is_intuBag}
\alias{intuEnv}
\alias{set_intuEnv}
\title{Experimental features.}
\description{
  Extensions (experimental) to pipelines. Please see vignette for examples of use.}
\usage{
as_intuBag(object)
clear_intuEnv()
intuBag(...)
is_intuBag(object)
intuEnv(...)
set_intuEnv(envir)
}
\arguments{
  \item{object}{Object.}
  \item{envir}{Environment.}
  \item{...}{Objects.}
}
\value{
  Return value of functions.
}
\author{
Roberto Bertolusso
}
\examples{
## Please see vignette
}
\keyword{intubate}
\keyword{magrittr}
